requirements.txt
openstack-heat
vcert
fabric
python-heatclient
keystoneauth1
six